fix: address code review issues for spaces phase 3
- Fix migration FK reference: users.id → user.id (would fail at runtime)
- Remove jsonb default from Column decorator (prevents schema drift)
- Add oldRole to member_role_change activity data (spec compliance)
- Fix double-fire of loadActivities on space refresh (track initialized ID)
- Extract formatTimeAgo to shared utility (was duplicated in 2 components)
